+++ backendPackage/.gitignore
@@ -1,1 +1,7 @@
 /ClaseCaca.class
+/VerificadorParentesis.class
+/BackendMain.class
+/InfixRPN.class
+/EvaluarRPN.class
+/Separador.class
+/Identificador.class
