+++ backendPackage/BackendMain.java
@@ -1,0 +1,33 @@
+package backendPackage;
+
+
+
+public class BackendMain {
+	
+	public static void main()
+	{
+		
+		String exp = "(((( 2 ^ 3 ) * 10)+5 - 27)/2/2 )*3";
+	    
+	    if (VerificadorParentesis.verificaParentesis(exp))
+	    {
+	        String[] expresion = Separador.separar(exp); 
+	        String[] output = InfixRPN.infixRPN(expresion);
+	        
+	        for (String token : output) {  
+	            System.out.print(token + " ");  
+	        }  
+	        
+	        double resultadoFinal = EvaluarRPN.evaluarRPN(output);
+	        
+	        System.out.println("\n" + resultadoFinal);
+	    }
+	    else
+	    {
+	        System.out.println("Expresion no tiene misma cantidad de parentesis izquierdos y derechos");
+	    }
+	}
+	
+   
+	
+}
