--- backendPackage/ClaseCaca.java
@@ -1,5 +1,0 @@
-package backendPackage;
-
-public class ClaseCaca {
-
-}
