+++ frontendPackage/MainGUI.java
@@ -1,5 +1,6 @@
 package frontendPackage;
 
+
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Dimension;
@@ -23,8 +24,12 @@ import javax.swing.JTextArea;
 import javax.swing.JTextField;
 import javax.swing.SwingUtilities;
 import javax.swing.UIManager;
-import java.util.ArrayList;
-import java.util.Stack;
+
+import backendPackage.BackendMain;
+
+//import backendPackage.VerificadorParentesis;
+//import java.util.ArrayList;
+//import java.util.Stack;
 //import java.util.regex.*;
 //import java.util.*;
 
@@ -43,29 +48,10 @@ public class MainGUI {
     JLabel 		picLabel;
 
     public static void main(String[] args) {
-      
     	
-     String exp = "(((( 2 ^ 3 ) * 10)+5 - 27)/2/2 )*3";
-        
-        
-        if (verificadorParentesis(exp))
-        {
-            String[] expresion = separador(exp); 
-            String[] output = infixRPN(expresion);
-            
-            for (String token : output) {  
-                System.out.print(token + " ");  
-            }  
-            
-            double resultadoFinal = evaluarRPN(output);
-            
-            System.out.println("\n" + resultadoFinal);
-        }
-        else
-        {
-            System.out.println("Expresion no tiene misma cantidad de parentesis izquierdos y derechos");
-        }
-
+    	
+      BackendMain.main();
+    	
     	
         SwingUtilities.invokeLater(new Runnable() {
             @Override
@@ -81,173 +67,8 @@ public class MainGUI {
             }
         });
     }
-   
-    public static String[] infixRPN(String[] pedazos) //Basado en http://www.technical-recipes.com/2011/a-mathematical-expression-parser-in-java-and-cpp/
-    {
-        ArrayList<String> out = new ArrayList<String>();
-        Stack<String> stack = new Stack<String>();
-            
-        for (String pedazo:pedazos)    //foreach sobre pedazos
-        {
-            if (esNumero(pedazo))    // numero
-            {
-                out.add(pedazo);
-            }
-            else if(pedazo.equals("("))    //parentesis izquierdo
-            {
-                stack.push(pedazo);
-            }
-            else if(pedazo.equals(")"))    //parentesis derecho
-            {
-                while(!stack.empty() && !stack.peek().equals("("))
-                {
-                    out.add(stack.pop());
-                }
-                stack.pop();
-            }
-            
-            else if(esOperador(pedazo))// operador
-            {
-                while (!stack.empty() && esOperador(stack.peek()))   
-                {                      
-                       out.add(stack.pop());     
-                }  
-                stack.push(pedazo);    // Push al nuevo operador del stack  
-            }
-        }
-        while(!stack.empty())
-        {
-            out.add(stack.pop());       
-        }
-        String[] output = new String[out.size()];  
-        return out.toArray(output);  
-    }
-    
-    public static double evaluarRPN(String[] pedazos) // Basado en http://www.technical-recipes.com/2011/a-mathematical-expression-parser-in-java-and-cpp/
-    {
-        Stack<String> stack = new Stack<String>();
-        
-        for (String pedazo : pedazos)
-        {
-            if (esNumero(pedazo))
-            {
-                stack.push(pedazo);
-            } else if(esOperador(pedazo))
-            {
-                double primero = Double.parseDouble(stack.pop().replace(",", "."));
-                double segundo = Double.parseDouble(stack.pop().replace(",", "."));
-                
-                double resultado = 0;
-                
-                
-                if (pedazo.equals("+"))
-                { resultado = primero + segundo; }
-                else if (pedazo.equals("-"))
-                { resultado = segundo - primero; }
-                else if (pedazo.equals("*"))
-                { resultado = primero * segundo; }
-                else if (pedazo.equals("/"))
-                { resultado = segundo / primero; }
-                else if (pedazo.equals("^"))
-                { resultado = Math.pow(segundo, primero); }
-                
-                stack.push(String.valueOf(resultado));
-            }
-        }
-        return Double.parseDouble(stack.pop());
-    }
-    
-    public static boolean esNumero(String numero) // De: http://stackoverflow.com/questions/11241690/regex-for-checking-if-a-string-is-strictly-alphanumeric
-    {                                                // y: http://stackoverflow.com/questions/21704583/regex-for-documentfilter-to-match-all-decimal-number-but-also-number-with-just-a
-        String pattern= "^[0-9]+[.]?[0-9]{0,}$"; // "^[0-9]*$";
-        if(numero.matches(pattern)){
-            return true;
-        }
-        return false;  
-    }
-    
-    public static boolean esOperador(String numero)
 
-    {
-        if (numero.equals("+") || numero.equals("-") ||
-                numero.equals("*") || numero.equals("/")
-                || numero.equals("^"))
-        {
-            return true;
-        } else
-        {
-            return false;
-        }
-    }
 
-    public static String[] separador(String expresion)
-    {
-        expresion = expresion.replace(" ", "");
-        ArrayList<String> lista = new ArrayList<String>();
-        
-        int i = 0;
-        
-        String aux = "";
-        
-        while (i < expresion.length())
-        {
-                if(esOperador(Character.toString(expresion.charAt(i)))
-                        || Character.toString(expresion.charAt(i)).equals("(")
-                        || Character.toString(expresion.charAt(i)).equals( ")"))
-                {        
-                    if (!aux.equals(""))
-                    {
-                        lista.add(aux);                                        //Agregar todo lo que estaba antes del operador o parentesis
-                        aux = "";                                            // Reinicializar aux
-                    }
-                    
-                    lista.add(Character.toString(expresion.charAt(i)));    //Agregar operador
-                } 
-                else // Si es que es un numero
-                {
-                    aux = aux + Character.toString(expresion.charAt(i));
-                }
-                i++;
-        }
-        
-        if (!aux.equals(""))
-        {
-            lista.add(aux);
-        }
-        
-        String[] output = new String[lista.size()];
-        return lista.toArray(output);
-    }
-
-    public static boolean verificadorParentesis(String expresion)
-    {
-        int i = 0;
-        int izq = 0;
-        int der = 0;
-        
-        while (i < expresion.length())
-        {
-            if (Character.toString(expresion.charAt(i)).equals("("))
-            {
-                izq++;
-            } 
-            else if (Character.toString(expresion.charAt(i)).equals(")"))
-            {
-                der++;
-            }
-            i++;
-        }
-        
-        if (izq == der)
-        {
-            return true;
-        }
-        else
-        {
-            return false;
-        }
-        
-    }
     
     BufferedImage logo;
     public void preDisplay() {
@@ -358,7 +179,7 @@ public class MainGUI {
         subPanel.add(boton2);
         subPanel.add(boton3);
 
-        //Now we simply add it to your main panel.
+        
         panelCentral.add(subPanel, BorderLayout.CENTER);
